requirements.txt

psycopg2-binary